+++ docs/quick_tutorial/cookiecutters/.coveragerc
@@ -1,0 +1,3 @@
+[run]
+source = cc_starter
+omit = cc_starter/test*
