+++ docs/quick_tutorial/cookiecutters/CHANGES.txt
@@ -1,0 +1,4 @@
+0.0
+---
+
+- Initial version.
