+++ docs/quick_tutorial/cookiecutters/MANIFEST.in
@@ -1,0 +1,2 @@
+include *.txt *.ini *.cfg *.rst
+recursive-include cc_starter *.ico *.png *.css *.gif *.jpg *.pt *.txt *.mak *.mako *.js *.html *.xml *.jinja2
