+++ docs/quick_tutorial/cookiecutters/cc_starter/views.py
@@ -1,0 +1,6 @@
+from pyramid.view import view_config
+
+
+@view_config(route_name='home', renderer='templates/mytemplate.jinja2')
+def my_view(request):
+    return {'project': 'cc_starter'}
