+++ docs/quick_tutorial/cookiecutters/development.ini
@@ -4,14 +4,14 @@
 ###
 
 [app:main]
-use = egg:scaffolds
+use = egg:cc_starter
 
 pyramid.reload_templates = true
 pyramid.debug_authorization = false
 pyramid.debug_notfound = false
 pyramid.debug_routematch = false
 pyramid.default_locale_name = en
-pyramid.includes = 
+pyramid.includes =
     pyramid_debugtoolbar
 
 # By default, the toolbar only appears for clients from IP addresses
@@ -24,7 +24,7 @@ pyramid.includes =
 
 [server:main]
 use = egg:waitress#main
-listen = *:6543
+listen = 127.0.0.1:6543 [::1]:6543
 
 ###
 # logging configuration
@@ -32,7 +32,7 @@ listen = *:6543
 ###
 
 [loggers]
-keys = root, scaffolds
+keys = root, cc_starter
 
 [handlers]
 keys = console
@@ -44,10 +44,10 @@ keys = generic
 level = INFO
 handlers = console
 
-[logger_scaffolds]
+[logger_cc_starter]
 level = DEBUG
 handlers =
-qualname = scaffolds
+qualname = cc_starter
 
 [handler_console]
 class = StreamHandler
@@ -56,4 +56,4 @@ level = NOTSET
 formatter = generic
 
 [formatter_generic]
-format = %(asctime)s %(levelname)-5.5s [%(name)s][%(threadName)s] %(message)s
+format = %(asctime)s %(levelname)-5.5s [%(name)s:%(lineno)s][%(threadName)s] %(message)s
