+++ docs/quick_tutorial/cookiecutters/production.ini
@@ -4,7 +4,7 @@
 ###
 
 [app:main]
-use = egg:scaffolds
+use = egg:cc_starter
 
 pyramid.reload_templates = false
 pyramid.debug_authorization = false
@@ -26,7 +26,7 @@ listen = *:6543
 ###
 
 [loggers]
-keys = root, scaffolds
+keys = root, cc_starter
 
 [handlers]
 keys = console
@@ -38,10 +38,10 @@ keys = generic
 level = WARN
 handlers = console
 
-[logger_scaffolds]
+[logger_cc_starter]
 level = WARN
 handlers =
-qualname = scaffolds
+qualname = cc_starter
 
 [handler_console]
 class = StreamHandler
@@ -50,4 +50,4 @@ level = NOTSET
 formatter = generic
 
 [formatter_generic]
-format = %(asctime)s %(levelname)-5.5s [%(name)s][%(threadName)s] %(message)s
+format = %(asctime)s %(levelname)-5.5s [%(name)s:%(lineno)s][%(threadName)s] %(message)s
