+++ docs/quick_tutorial/cookiecutters/pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+testpaths = cc_starter
+python_files = *.py
