--- docs/quick_tutorial/scaffolds/CHANGES.txt
@@ -1,4 +1,0 @@
-0.0
----
-
--  Initial version
