--- docs/quick_tutorial/scaffolds/MANIFEST.in
@@ -1,2 +1,0 @@
-include *.txt *.ini *.cfg *.rst
-recursive-include scaffolds *.ico *.png *.css *.gif *.jpg *.pt *.txt *.mak *.mako *.js *.html *.xml
