--- docs/quick_tutorial/scaffolds/README.txt
@@ -1,1 +1,0 @@
-scaffolds README
