--- docs/quick_tutorial/scaffolds/scaffolds/views.py
@@ -1,6 +1,0 @@
-from pyramid.view import view_config
-
-
-@view_config(route_name='home', renderer='templates/mytemplate.pt')
-def my_view(request):
-    return {'project': 'scaffolds'}
